make dtn dimer apart of the ecosystem
